+++ src/main/java/racingGame/controller/Main.java
@@ -8,7 +8,6 @@ public class Main {
     public static void main(String[] args) {
         RacingGame racingGame = new RacingGame();
         racingGame.settingCar(InputView.inputCar());
-        racingGame.tryGame(InputView.inputTry());
-        ResultView.printResult(racingGame.getCars());
+        ResultView.printResult(racingGame.tryAllCarGame(InputView.inputTry()));
     }
 }
