+++ src/main/java/racingGame/model/RacingGame.java
@@ -1,10 +1,13 @@
 package racingGame.model;
 
+import racingGame.util.RandomGenerator;
+
 import java.util.ArrayList;
 import java.util.List;
 
 public class RacingGame {
     private List<Car> cars;
+    private static int MOVE_RESTRICTION = 4;
 
     public void settingCar(int numCar) {
         cars = new ArrayList<>();
@@ -14,13 +17,29 @@ public class RacingGame {
         }
     }
 
-    public void tryGame(int numberOfTry) {
+    public List<Car> tryAllCarGame(int numberOfTry) {
         for (Car car : cars) {
-            car.tryGame(numberOfTry);
+            tryEachCarGame(car, numberOfTry);
+        }
+        return cars;
+    }
+
+    private void tryEachCarGame(Car car, int numberOfTry) {
+        for (int i = 0; i < numberOfTry; i++) {
+            move(car);
         }
     }
 
-    public List<Car> getCars() {
-        return cars;
+    private void move(Car car) {
+        if (isMove()) {
+            car.carMove();
+        }
+    }
+
+    private boolean isMove() {
+        if (RandomGenerator.randomGenerator() >= MOVE_RESTRICTION) {
+            return true;
+        }
+        return false;
     }
 }
